pyproject.toml
[project]
name = "malaga-tracker"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "amadeus>=9.0",
    "python-dotenv>=1.0",
    "requests>=2.31",
]
